- reworked Flickr plugin to not include Flickr API key    - prepared Flickr plugin for cheeseshop upload
git-svn-id: http://svn.osafoundation.org/chandler/trunk@13580 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/projects/Chandler-FlickrPlugin/flickr/__init__.py
@@ -1,4 +1,4 @@
-#   Copyright (c) 2005-2006 Open Source Applications Foundation
+#   Copyright (c) 2005-2007 Open Source Applications Foundation
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -12,26 +12,22 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+import flickr, wx, dateutil
 
-import application.dialogs.Util
+from application.dialogs.Util import promptUser
+from dialogs import LicenseTask, promptLicense
 from application import schema
-import flickr
-from flickr import FlickrError, FlickrNotFoundError
 from osaf import pim
 from photos import PhotoMixin
-import osaf.views.detail as Detail
 from osaf.pim.collections import KindCollection
 from repository.util.URL import URL
 from repository.item.Item import MissingClass
 from datetime import datetime
-import dateutil
-import wx
-import logging
 from i18n import MessageFactory
 from osaf import messages
 from osaf.pim.structs import SizeType, RectType
-from osaf.framework.blocks.Block import *
-from osaf.framework.blocks.MenusAndToolbars import MenuItem
+from osaf.framework.blocks import BlockEvent, MenuItem, AddToSidebarEvent
+from osaf.framework.blocks.Block import Block
 from osaf.startup import PeriodicTask
 from datetime import timedelta
 from osaf.usercollections import UserCollection
@@ -39,7 +35,6 @@ from PyICU import ICUtzinfo
 
 _ = MessageFactory("Chandler-FlickrPlugin")
 
-logger = logging.getLogger(__name__)
 
 class FlickrPhotoMixin(PhotoMixin):
     """
@@ -57,25 +52,26 @@ class FlickrPhotoMixin(PhotoMixin):
         )
     )
 
-    def __init__(self, photo=None,*args,**kwargs):
-        super(FlickrPhotoMixin,self).__init__(*args,**kwargs)
+    def __init__(self, photo=None, *args, **kwds):
+
+        super(FlickrPhotoMixin,self).__init__(*args,**kwds)
+
         if photo is not None:
             self.flickrID = photo.id
             self.displayName = photo.title
             self.description = photo.description.encode('utf8')
             self.owner = photo.owner.username
-            if photo.owner.realname is not None and len(photo.owner.realname.strip()) > 0:
+            if (photo.owner.realname is not None and
+                photo.owner.realname.strip()):
                 self.owner = photo.owner.realname
     
             self.imageURL = URL(photo.getURL(urlType="source"))
             self.datePosted = datetime.utcfromtimestamp(int(photo.dateposted))
             self.dateTaken = dateutil.parser.parse(photo.datetaken, 
-                                default=datetime.now(tz=ICUtzinfo.default))
-            try:
-                if photo.tags:
-                    self.tags = [Tag.getTag(self.itsView, tag.text) for tag in photo.tags]
-            except Exception, e:
-                logging.exception(e)
+                                                   default=datetime.now(tz=ICUtzinfo.default))
+            if photo.tags:
+                self.tags = [Tag.getTag(self.itsView, tag.text)
+                             for tag in photo.tags]
 
         self.importFromURL(self.imageURL)
         
@@ -87,21 +83,25 @@ class FlickrPhotoMixin(PhotoMixin):
         super(FlickrPhotoMixin, self).addDisplayWhos(whos)
         owner = getattr(self, 'owner', None)
         if owner is not None:
-            whos.append((10, owner, 'owner'))
+            whos.append((16, owner, 'owner'))
+
 
 class FlickrPhoto(FlickrPhotoMixin, pim.Note):
     pass
 
+
 class Tag(pim.ContentItem):
     """
-    Tag are items with bidirectional references to all the FlickrPhoto's with that
-    tag. This makes it easy to find all photos with a given tag or all tags belonging
-    to a photo. Currently, there isn't any code that takes advantage of Tags.
+    Tag are items with bidirectional references to all the FlickrPhoto's
+    with that tag. This makes it easy to find all photos with a given tag or
+    all tags belonging to a photo. Currently, there isn't any code that
+    takes advantage of Tags.
     """
-    itemsWithTag = schema.Sequence(FlickrPhotoMixin, inverse=FlickrPhotoMixin.tags)
+    itemsWithTag = schema.Sequence(FlickrPhotoMixin,
+                                   inverse=FlickrPhotoMixin.tags)
 
     @classmethod
-    def getTag (cls, view, tagName):
+    def getTag(cls, view, tagName):
         """
         Factory Method for getting a Tag.
 
@@ -127,26 +127,31 @@ class Tag(pim.ContentItem):
         # make a new Tag
         newTag = Tag(itsView=view)
         newTag.displayName = tagName
+
         return newTag
+
 
 class PhotoCollection(pim.ListCollection):
     """
     A ListCollection of FlickrPhotos
     """
-    userName = schema.One (schema.Text, initialValue=u'')
-    tag = schema.One (Tag, initialValue=None)
-    fillInBackground = schema.One (schema.Boolean, defaultValue=False)
+    userName = schema.One(schema.Text, initialValue=u'')
+    tag = schema.One(Tag, initialValue=None)
+    fillInBackground = schema.One(schema.Boolean, defaultValue=False)
 
-
-    def fillCollectionFromFlickr(self, repView):
+    def fillCollectionFromFlickr(self, view, n=16, apiKey=None):
         """
         Fills the collection with photos from the flickr website.
         """
+        if apiKey:
+            flickr.setLicense(apiKey)
+
         if self.userName:
             flickrUserName = flickr.people_findByUsername(self.userName.encode('utf8'))
-            flickrPhotos = flickr.people_getPublicPhotos(flickrUserName.id,10)
+            flickrPhotos = flickr.people_getPublicPhotos(flickrUserName.id, n)
         elif self.tag:
-            flickrPhotos = flickr.photos_search(tags=self.tag,per_page=10,sort="date-posted-desc")
+            flickrPhotos = flickr.photos_search(tags=self.tag, per_page=n,
+                                                sort="date-posted-desc")
         else:
             assert(False, "we should have either a userName or tag")
 
@@ -154,7 +159,7 @@ class PhotoCollection(pim.ListCollection):
         # an index named flickerIDIndex which indexes all the photos by
         # their flickrID which makes it easy to quickly lookup any photo by
         # index.
-        flickrPhotosCollection = schema.ns('flickr', repView).flickrPhotosCollection
+        flickrPhotosCollection = schema.ns('flickr', view).flickrPhotosCollection
         for flickrPhoto in flickrPhotos:
             """
             If we've already downloaded a photo with this id use it instead.
@@ -164,15 +169,17 @@ class PhotoCollection(pim.ListCollection):
                 'exact',         # require an exact match
                                  # compare function
                 lambda uuid: cmp(flickrPhoto.id,
-                                 repView.findValue(uuid, 'flickrID')))
+                                 view.findValue(uuid, 'flickrID')))
 
             if photoUUID is None:
-                photoItem = FlickrPhoto(photo=flickrPhoto, itsView=repView)
+                photoItem = FlickrPhoto(photo=flickrPhoto, itsView=view)
             else:
-                photoItem = repView[photoUUID]
+                photoItem = view[photoUUID]
 
-            self.add (photoItem)
-        repView.commit()
+            self.add(photoItem)
+
+        view.commit()
+
 
 class UpdateTask(object):
     """
@@ -182,81 +189,96 @@ class UpdateTask(object):
         self.view = item.itsView.repository.createView("Flickr")
 
     def run(self):
-        logger.info("receiveWakeupCall()")
-
         # We need the view for most repository operations
         self.view.refresh(notify=False)
 
         # Go through all the PhotoCollections and update those that
         # have fillInBackground set. fillCollectionFromFlickr commits
-        for myPhotoCollection in PhotoCollection.iterItems(self.view):
-            if myPhotoCollection.fillInBackground:
-                myPhotoCollection.fillCollectionFromFlickr(self.view)
+        for collection in PhotoCollection.iterItems(self.view):
+            if collection.fillInBackground:
+                collection.fillCollectionFromFlickr(self.view)
 
         return True
 
+
 class CollectionTypeEnumType(schema.Enumeration):
     """
-    An enumeration used to specify two different kinds of AddFlickrCollectionEvent
-    types.
+    An enumeration used to specify two different kinds of
+    AddFlickrCollectionEvent types.
     """
     values = "Tag", "Owner"
+
 
 class AddFlickrCollectionEvent(AddToSidebarEvent):
     """
     An event used to add a new FlickrCollection to the sidebar.
     """
-    collectionType = schema.One (CollectionTypeEnumType, initialValue = 'Tag')
+    collectionType = schema.One(CollectionTypeEnumType, initialValue='Tag')
     
-    def onNewItem (self):
+    def onNewItem(self):
         """
         Called to create a new collection that gets added to the sidebar.
         """
         photoCollection = None
         if self.collectionType == 'Owner':
-            userName = application.dialogs.Util.promptUser(
-                messages.USERNAME,
-                _(u"Enter a Flickr user name"))
+            userName = promptUser(messages.USERNAME,
+                                  _(u"Enter a Flickr user name"))
             if userName is not None:
-                photoCollection = PhotoCollection (itsView = self.itsView)
+                photoCollection = PhotoCollection(itsView=self.itsView)
                 photoCollection.userName = userName
                 photoCollection.displayName = userName
         else:
             assert (self.collectionType == 'Tag')
-            tagString = application.dialogs.Util.promptUser(
-                _(u"Tag"),
-                _(u"Enter a Flickr Tag"))
+            tagString = promptUser(_(u"Tag"),
+                                   _(u"Enter a Flickr Tag"))
             if tagString is not None:
-                photoCollection = PhotoCollection (itsView = self.itsView)
+                photoCollection = PhotoCollection(itsView=self.itsView)
                 photoCollection.tag = Tag.getTag(self.itsView, tagString)
                 photoCollection.displayName = photoCollection.tag.displayName
 
         if photoCollection is not None:
-            # Setting perferredKind to None will cause it to be displayed in the All View
-            UserCollection (photoCollection).preferredClass = MissingClass
-            try:
-                photoCollection.fillCollectionFromFlickr(self.itsView)
-            except flickr.FlickrError, flickrException:
-                wx.MessageBox (unicode(flickrException))
-                photoCollection.delete()
-                photoCollection = None
-            else:
-                photoCollection.fillInBackground = True
+            # Setting preferredKind to None will cause it to be displayed in
+            # the All View 
+            UserCollection(photoCollection).preferredClass = MissingClass
+            while True:
+                try:
+                    photoCollection.fillCollectionFromFlickr(self.itsView)
+                    photoCollection.fillInBackground = True
+                    break
+                except flickr.FlickrError, e:
+                    if "api key" in e.args[0].lower():
+                        if promptLicense():
+                            continue
+                    wx.MessageBox(unicode(e))
+                    photoCollection.delete()
+                    photoCollection = None
+                    break
 
         return photoCollection
 
+
+class FlickrHandler(Block):
+
+    def on_flickr_UpdateFlickrNowEvent(self, event):
+        schema.ns('flickr', self.itsView).FlickrUpdateTask.run_once()
+
+
 def installParcel(parcel, oldVersion=None):
     """
-    Creates Items that live in the initial repository. Run at repository buildtime.
+    Creates Items that live in the initial repository. 
+
+    Run at repository buildtime. 
     """
-    # A KindCollection of all FlickPhoto kinds. The flickrIDIndex is used to lookup
-    # photos by flickrID quickly.
+
+    # A KindCollection of all FlickPhoto kinds. The flickrIDIndex is used to
+    # lookup photos by flickrID quickly.
     flickrPhotosCollection = KindCollection.update(
         parcel, 'flickrPhotosCollection',
         kind = FlickrPhotoMixin.getKind(parcel.itsView),
         recursive = True)
 
-    flickrPhotosCollection.addIndex ('flickrIDIndex', 'attribute', attribute='flickrID', compare="__cmp__")
+    flickrPhotosCollection.addIndex('flickrIDIndex', 'attribute',
+                                    attribute='flickrID', compare="__cmp__")
 
     # A NewFlickrCollectionEvent that adds a "Owner" collection to the sidebar
     addFlickrCollectionByOwnerEvent = AddFlickrCollectionEvent.update(
@@ -269,6 +291,16 @@ def installParcel(parcel, oldVersion=None):
         parcel, 'addFlickrCollectionByTagEvent',
         blockName = 'addFlickrCollectionByTagEvent',
         collectionType = 'Tag')
+
+    handler = FlickrHandler.update(parcel, 'flickrHandler',
+                                   blockName='_flickr_flickrHandler')
+
+    # A BlockEvent that tickles the background update task to run
+    # Add an event for managing p2p access to collections
+    updateEvent = BlockEvent.update(parcel, None,
+                                    blockName='_flickr_UpdateFlickrNow',
+                                    dispatchEnum='SendToBlockByReference',
+                                    destinationBlockReference=handler)
 
     # Add menu items to Chandler
     collectionMenu = schema.ns('osaf.views.main', parcel).ExperimentalMenu
@@ -289,20 +321,33 @@ def installParcel(parcel, oldVersion=None):
  
     MenuItem.update(
         parcel, 'NewFlickrCollectionByTag',
-        blockName = 'NewFlickrCollectionByTagIMenutem',
+        blockName = 'NewFlickrCollectionByTagMenuItem',
         title = _(u'New Flickr Collection by Tag'),
         event = addFlickrCollectionByTagEvent,
         eventsForNamedLookup = [addFlickrCollectionByTagEvent],
         parentBlock = collectionMenu)
 
+    MenuItem.update(
+        parcel, 'UpdateFlickrNow',
+        blockName = 'UpdateFlickrNowMenuItem',
+        title = _(u'Update Photo Collections'),
+        event = updateEvent,
+        eventsForNamedLookup = [updateEvent],
+        parentBlock = collectionMenu)
 
     # The periodic task that adds new photos to the collection in the background
     PeriodicTask.update(
         parcel, 'FlickrUpdateTask',
         invoke = 'flickr.UpdateTask',
         run_at_startup = True,
-        interval = timedelta(minutes=2))
+        interval = timedelta(minutes=60))
 
+    # The periodic task that sets the Flickr's API key
+    PeriodicTask.update(parcel, "licenseTask",
+                        invoke="flickr.LicenseTask",
+                        interval=timedelta(days=1),
+                        run_at_startup=True)
+    LicenseTask(None).run()
 
     # The detail view used to display a flickrPhoto
     blocks = schema.ns('osaf.framework.blocks', parcel)
